Restricted login, edit and delete room pages

+++ base/views.py
@@ -1,6 +1,8 @@
+from django.http import HttpResponse
 from django.shortcuts import render, redirect
 from django.contrib.auth.models import User
 from django.contrib.auth import authenticate, login, logout
+from django.contrib.auth.decorators import login_required
 from django.contrib import messages
 from django.db.models import Q
 from .models import Room, Topic
@@ -8,6 +10,9 @@ from .forms import RoomForm
 
 
 def loginView(request):
+
+    if request.user.is_authenticated:
+        return redirect('home')
 
     if request.method == 'POST':
         username = request.POST.get('username')
@@ -31,6 +36,10 @@ def loginView(request):
     context = {}
     return render(request, 'base/login_register.html', context)
 
+def logoutView(request):
+    logout(request)
+    return redirect('home')
+
 def home(request):
     q = request.GET.get('q') if request.GET.get('q') != None else ''
     rooms = Room.objects.filter(
@@ -49,6 +58,7 @@ def room(request, pk):
     context = {'room': room}
     return render(request, 'base/room.html', context)
 
+@login_required(login_url='login')
 def createRoom(request):
     form = RoomForm()
 
@@ -63,10 +73,13 @@ def createRoom(request):
     context = {'form': form}
     return render(request, 'base/room_form.html', context)
 
-
+@login_required(login_url='login')
 def updateRoom(request, pk):
     room = Room.objects.get(id=pk)
     form = RoomForm(instance=room)
+
+    if(request.user != room.host):
+        return HttpResponse("You are not allowed here!!!")
 
     if(request.method == 'POST'):
         form = RoomForm(request.POST, instance=room)
@@ -77,8 +90,12 @@ def updateRoom(request, pk):
     context = {'form': form}
     return render(request, 'base/room_form.html', context)
 
+@login_required(login_url='login')
 def deleteRoom(request, pk):
     room = Room.objects.get(id=pk)
+
+    if(request.user != room.host):
+        return HttpResponse("You are not allowed here!!!")
 
     if(request.method == 'POST'):
         room.delete()
